+++ backend/apps/_archived/account/templates/account/mail/activation_request.html
