+++ backend/apps/_archived/account/templates/account/registration.html
