--- backend/apps/exchange/templates/exchange/add_exchange.html
@@ -1,15 +1,0 @@
-{% extends 'base.html' %}
-{% load crispy_forms_tags %}
-{% block content%}
-<br \>
-<div style="clear:left">
-    <h2>Proposer un échange sur la Bourse aux options</h2>
-    </br>
-    <form method='POST'>
-        {% csrf_token %}
-        {{form|crispy}}
-        </br>
-        <button class='btn btn-dark'>Ajouter ma proposition d'échange</button>
-    </form>
-</div>
-{% endblock %}
