--- backend/apps/exchange/templates/exchange/delete_exchange.html
@@ -1,14 +1,0 @@
-{% extends 'base.html' %}
-{% load crispy_forms_tags %}
-{% block content%}
-<br \>
-<div style="clear:left">
-    <h2>Supprimer un échange sur la Bourse aux options</h2>
-    </br>
-    <form method="post">
-        {% csrf_token %}
-        <p>Voulez-vous vraiment supprimer votre échange ?</p>
-        <button class='btn btn-dark'>Supprimer</button>
-    </form>
-</div>
-{% endblock %}
