--- backend/apps/exchange/templates/exchange/exchange.html
@@ -1,49 +1,0 @@
-{% extends 'base.html' %}
-{% block content%}
-<br/>
-<div style="clear:left">
-    <h2>Bourse aux options</h2>
-    <div>
-    <div class="card-columns" style="width: 50rem; margin: 0 auto;">
-      {% if not request.user.student.exchange %} 
-      <div class="card bg-secondary" style="width: 15rem; box-shadow: 1px 1px 7px #999">
-        <div class="card-body">
-          <h5 class="card-title" style="color: white">Ça m'intéresse !</h5>
-          <p class="card-text" style="color: white">Pas satisfait.e de l'option que vous avez obtenue ? Proposez un échange ! </p>
-          <a href="{% url 'exchange:add-exchange'%}" class="btn btn-dark">Je propose un échange</a>
-        </div>
-      </div>
-      {% endif %}
-      {% for exchange in exchanges %}
-        <div class="card" style="width: 15rem; box-shadow: 1px 1px 7px #999">
-          <img src={{exchange.got.image}} class="card-img-top" alt={{exchange.got}} style="height: 100px;">
-          <div class="card-body py-3">
-            <h5 class="card-title">{{exchange.student.name }}</h5>
-            <p class="card-text">Option proposée :</p>
-          </div>
-          <ul class="list-group list-group-flush">
-            <li class="list-group-item py-1" style="background-color: rgb(236, 236, 236); margin-top: -10px;">{{exchange.got}}</li>
-          </ul>
-          <div class="card-body py-1">
-            <p class="card-text">{% if exchange.wanted.all|length == 1 %}Option recherchée :{% else %}Options recherchées : {% endif %}</p>
-          </div>
-          <ul class="list-group list-group-flush">
-            {% for option in exchange.wanted.all %} 
-            <li class="list-group-item py-1" style="background-color: rgb(236, 236, 236);">{{option}}</li>
-            {% endfor %}
-          </ul>
-          <div class="card-body py-3">
-            {% if exchange.student.id == request.user.student.id %}
-            <a href="{% url 'exchange:update-exchange' exchange.id %}" class="btn btn-dark">Modifier</a>
-            <a href="{% url 'exchange:delete-exchange' exchange.id %}" class="btn btn-dark">Supprimer</a>
-            {% else %}
-            <a href="{% url 'student:detail' exchange.student.id %}" class="btn btn-dark">Contacter</a>
-            {% endif %}
-          </div>
-        </div>
-      <br/>
-      {% endfor %}
-    </div>
-</div>
-
-{% endblock%}
