--- backend/apps/exchange/templates/exchange/update_exchange.html
@@ -1,15 +1,0 @@
-{% extends 'base.html' %}
-{% load crispy_forms_tags %}
-{% block content%}
-<br \>
-<div style="clear:left">
-    <h2>Modifier un échange sur la Bourse aux options</h2>
-    </br>
-    <form method='POST'>
-        {% csrf_token %}
-        {{form|crispy}}
-        </br>
-        <button class='btn btn-dark'>Modifier</button>
-    </form>
-</div>
-{% endblock %}
