+++ backend/config/settings/docker.py
@@ -82,11 +82,11 @@ CELERY_TIMEZONE = 'Europe/Paris'
 # make the cron tasks be executed every monday at 4am
 CELERY_BEAT_SCHEDULE = {
     'remove-inactive-accounts': {
-        'task': 'apps.account.tasks.remove_inactive_accounts',
+        'task': 'apps._archived.account.tasks.remove_inactive_accounts',
         'schedule': crontab(minute=0, hour=4, day_of_week=1)
     },
     'remove-temp-access': {
-        'task': 'apps.account.tasks.remove_temporary_access',
+        'task': 'apps._archived.account.tasks.remove_temporary_access',
         'schedule': crontab(minute=0, hour=4, day_of_week=1)
     },
     'clean-notifications': {
